+++ exchanges/kraken/kraken_test.go
@@ -748,166 +748,3 @@ func TestOrderBookOutOfOrder(t *testing.T) {
 		t.Error("Expected out of order orderbook error")
 	}
 }
-
-// TestSubscribeToChannel websocket test
-func TestSubscribeToChannel(t *testing.T) {
-	if k.Name == "" {
-		k.SetDefaults()
-		TestSetup(t)
-	}
-	if !k.Websocket.IsEnabled() {
-		t.Skip("Websocket not enabled, skipping")
-	}
-	if k.WebsocketConn == nil {
-		k.Websocket.Connect()
-	}
-
-	err := k.WsSubscribeToChannel("ticker", []string{"XTZ/USD"}, 1)
-	if err != nil {
-		t.Error(err)
-	}
-}
-
-// TestSubscribeToNonExistentChannel websocket test
-func TestSubscribeToNonExistentChannel(t *testing.T) {
-	if k.Name == "" {
-		k.SetDefaults()
-		TestSetup(t)
-	}
-	if !k.Websocket.IsEnabled() {
-		t.Skip("Websocket not enabled, skipping")
-	}
-	if k.WebsocketConn == nil {
-		k.Websocket.Connect()
-	}
-	err := k.WsSubscribeToChannel("ticker", []string{"pewdiepie"}, 1)
-	if err != nil {
-		t.Error(err)
-	}
-	subscriptionError := false
-	for i := 0; i < 7; i++ {
-		response := <-k.Websocket.DataHandler
-		if err, ok := response.(error); ok && err != nil {
-			subscriptionError = true
-			break
-		}
-	}
-	if !subscriptionError {
-		t.Error("Expected error")
-	}
-}
-
-// TestSubscribeUnsubscribeToChannel websocket test
-func TestSubscribeUnsubscribeToChannel(t *testing.T) {
-	if k.Name == "" {
-		k.SetDefaults()
-		TestSetup(t)
-	}
-	if !k.Websocket.IsEnabled() {
-		t.Skip("Websocket not enabled, skipping")
-	}
-	if k.WebsocketConn == nil {
-		k.Websocket.Connect()
-	}
-	err := k.WsSubscribeToChannel("ticker", []string{"XRP/JPY"}, 1)
-	if err != nil {
-		t.Error(err)
-	}
-	err = k.WsUnsubscribeToChannel("ticker", []string{"XRP/JPY"}, 2)
-	if err != nil {
-		t.Error(err)
-	}
-}
-
-// TestUnsubscribeWithoutSubscription websocket test
-func TestUnsubscribeWithoutSubscription(t *testing.T) {
-	if k.Name == "" {
-		k.SetDefaults()
-		TestSetup(t)
-	}
-	if !k.Websocket.IsEnabled() {
-		t.Skip("Websocket not enabled, skipping")
-	}
-	if k.WebsocketConn == nil {
-		k.Websocket.Connect()
-	}
-	err := k.WsUnsubscribeToChannel("ticker", []string{"QTUM/EUR"}, 3)
-	if err != nil {
-		t.Error(err)
-	}
-	unsubscriptionError := false
-	for i := 0; i < 5; i++ {
-		response := <-k.Websocket.DataHandler
-		if err, ok := response.(error); ok && err != nil {
-			if err.Error() == "requestID: '3'. Error: Subscription Not Found" {
-				unsubscriptionError = true
-				break
-			}
-		}
-	}
-	if !unsubscriptionError {
-		t.Error("Expected error")
-	}
-}
-
-// TestUnsubscribeWithChannelID websocket test
-func TestUnsubscribeWithChannelID(t *testing.T) {
-	if k.Name == "" {
-		k.SetDefaults()
-		TestSetup(t)
-	}
-	if !k.Websocket.IsEnabled() {
-		t.Skip("Websocket not enabled, skipping")
-	}
-	if k.WebsocketConn == nil {
-		k.Websocket.Connect()
-	}
-	err := k.WsUnsubscribeToChannelByChannelID(100)
-	if err != nil {
-		t.Error(err)
-	}
-	unsubscriptionError := false
-	for i := 0; i < 5; i++ {
-		response := <-k.Websocket.DataHandler
-		if err, ok := response.(error); ok && err != nil {
-			if err.Error() == "Not subscribed to the requested channelID" {
-				unsubscriptionError = true
-				break
-			}
-		}
-	}
-	if !unsubscriptionError {
-		t.Error("Expected error")
-	}
-}
-
-// TestUnsubscribeFromNonExistentChannel websocket test
-func TestUnsubscribeFromNonExistentChannel(t *testing.T) {
-	if k.Name == "" {
-		k.SetDefaults()
-		TestSetup(t)
-	}
-	if !k.Websocket.IsEnabled() {
-		t.Skip("Websocket not enabled, skipping")
-	}
-	if k.WebsocketConn == nil {
-		k.Websocket.Connect()
-	}
-	err := k.WsUnsubscribeToChannel("ticker", []string{"tseries"}, 0)
-	if err != nil {
-		t.Error(err)
-	}
-	unsubscriptionError := false
-	for i := 0; i < 5; i++ {
-		response := <-k.Websocket.DataHandler
-		if err, ok := response.(error); ok && err != nil {
-			if err.Error() == "Currency pair not in ISO 4217-A3 format tseries" {
-				unsubscriptionError = true
-				break
-			}
-		}
-	}
-	if !unsubscriptionError {
-		t.Error("Expected error")
-	}
-}
